requirements.txt
alembic==1.18.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.6.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
comtypes==1.4.14
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.1.7
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.0.0
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fqdn==1.5.1
fsspec==2026.1.0
greenlet==3.3.0
gunicorn
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
mysql-connector-python==9.6.0
mysqlclient==2.1.1
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
notebook==7.5.2
notebook_shim==0.2.4
numba==0.64.0
numpy==2.2.6
openai==2.16.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
platformdirs==4.5.1
pooch==1.9.0
preshed==3.0.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
PyPDF2==3.0.1
pypdfium2==5.5.0
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==2.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.4
SQLAlchemy==2.0.45
sqlparse==0.4.3
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
textblob==0.19.0
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2022.7
uri-template==1.3.0
urllib3==2.6.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wikipedia==1.4.0
wrapt==2.0.1
WTForms==3.1.2
